services/ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
redis==5.0.0
pydantic==2.9.0
structlog==24.4.0
python-dotenv==1.0.0
openai==1.45.0
